Organize SVEResult
